Add files via upload
Unlock the rhythm of Spotify's musical universe! 🎶 Dive into this data-driven journey, where Python meets Tableau, to explore the trends, rhythms, and beats that have shaped the music scene. From the Top 10 Artists to Genre Popularity Heatmaps, this repository is your backstage pass to a symphony of insights. Let the data play, and join the exploration! 🚀📊 #DataScience #MusicAnalysis #Spotify